se separaron algunas funciones del main.js para que no interrumpa con las otras paginas html que lo utilizan
