+++ destination.html
@@ -109,8 +109,8 @@
 					<h2>Viajá a tu lugar favorito</h2>
 				</div>
 			</div>
-			<div class="row">
-
+			<div class="row paquetes">
+<!-- 
 				<div class="col-lg-4 col-md-4 col-sm-6">
 					<a href="images/nuevayork.jpg" class="fh5co-card-item image-popup">
 						<figure>
@@ -192,7 +192,7 @@
 							<p><span class="btn btn-primary">Seleccionar</span></p>
 						</div>
 					</a>
-				</div>
+				</div> -->
 
 			</div>
 		</div>
@@ -329,6 +329,7 @@
 
 	<!-- Main -->
 	<script src="js/main.js"></script>
+	<script src="js/pintarPaquetes.js"></script>
 
 	</body>
 </html>
