+++ index.html
@@ -406,6 +406,8 @@
 
 	<!-- Main -->
 	<script src="js/main.js"></script>
+	<script src="js/signUpIndex.js"></script>
+	<script src="js/pintarPaquetes.js"></script>
 
 	</body>
 </html>
