+++ js/destinos.js
@@ -1,0 +1,3 @@
+import {pintarPaqTur} from './pintarPaquetes.js';
+
+pintarPaqTur();
